Split PasswordStorage into FileStorage and subclasses + Renamed Encrypter
PasswordStorage is now a child of FileStorage.

FileStorage is a base class which handles taking in a file name and has a virtual method of saving to file.

AccountStorage uses a map to read in and save accounts.

PasswordStorage simply saves passwords to the file.

Renamed encrypter to PasswordSecurity as class will be used for decrypting too.