Decryption Optimization
Optimized decryption method so that instead of generating the whole ascii collatz values, just returns the first one. If doesn't find a value, it returns -1.

This optimization will not be compatible with the program if need to return every possible character that could be in the position of the character within the password.